Moved out chain-signatures integration tests into own package
