+++ Cargo.toml
@@ -3,9 +3,6 @@ members = [
     # "mpc-recovery",
     "contract",
     "node",
-    "integration-tests",
-    # "integration-tests/chain-signatures",
-    # "integration-tests/mpc-recovery",
     # "load-tests",
     "keys",
     "test-oidc-provider",
