+++ integration-tests/chain-signatures/build.rs
@@ -1,0 +1,3 @@
+// HACK: need this build script so that env var OUT_DIR gets set:
+// https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-crates
+fn main() {}
